SmartHome/voltcraft2mqtt/requirements.txt
bluepy==1.3.0
paho_mqtt==2.1.0
python_json_logger==3.0.1
